requirements.txt
asgiref==3.7.2
Django==3.2.22
django-filter==23.3
djangorestframework==3.14.0
inflection==0.5.1
psutil==5.9.5
python-dotenv==0.21.1
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.7.1
